chore: break out all context hooks into hooks directory
