+++ src/components/common/RequireAuth.tsx
@@ -2,7 +2,7 @@ import { Redirect } from "raviger";
 import { PropsWithChildren } from "react";
 
 import { LoadingBackdrop } from "@/components";
-import { useAuth } from "@/context";
+import { useAuth } from "@/hooks";
 
 export const RequireAuth = ({ children }: PropsWithChildren<{}>) => {
   const { isAuthenticated, accessToken, isInitializing } = useAuth();
