+++ src/components/popovers/ProfileMiniCard.tsx
@@ -12,7 +12,7 @@ import {
 import { DateTime } from "luxon";
 import { navigate } from "raviger";
 
-import { useDeck, useMatch } from "@/context";
+import { useDeck, useMatch } from "@/hooks";
 import { getUserStats } from "@/logic";
 import { User } from "@/types";
