+++ src/components/toolbar/ThemeToggle.tsx
@@ -2,7 +2,7 @@ import Brightness4Icon from "@mui/icons-material/Brightness4";
 import Brightness7Icon from "@mui/icons-material/Brightness7";
 import { IconButton } from "@mui/material";
 
-import { useTheme } from "@/context";
+import { useTheme } from "@/hooks";
 
 export const ThemeToggle = () => {
   const { toggleTheme, mode } = useTheme();
