+++ src/hooks/index.ts
@@ -1,4 +1,11 @@
+export * from "./useApp";
+export * from "./useAuth";
+export * from "./useDeck";
 export * from "./useDecksFilters";
+export * from "./useFormat";
+export * from "./useMatch";
 export * from "./useMatchFilters";
 export * from "./usePersistentState";
+export * from "./useTheme";
+export * from "./useUser";
 export * from "./useUsersFilters";
