+++ src/hooks/useApp.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+
+import { AppContext } from "@/context";
+
+export const useApp = () => {
+  const ctx = useContext(AppContext);
+  if (!ctx) {
+    throw new Error("useApp must be used within an AppProvider");
+  }
+  return ctx;
+};
