+++ src/hooks/useAuth.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+
+import { AuthContext } from "@/context";
+
+export const useAuth = () => {
+  const ctx = useContext(AuthContext);
+  if (!ctx) {
+    throw new Error("useAuth must be used inside AuthProvider");
+  }
+  return ctx;
+};
