+++ src/hooks/useDeck.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+
+import { DeckContext } from "@/context";
+
+export const useDeck = () => {
+  const ctx = useContext(DeckContext);
+  if (!ctx) {
+    throw new Error("useDeck must be used within a DeckProvider");
+  }
+  return ctx;
+};
