+++ src/hooks/useUser.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+
+import { UserContext } from "@/context";
+
+export const useUser = () => {
+  const ctx = useContext(UserContext);
+  if (!ctx) {
+    throw new Error("useUser must be used within a UserProvider");
+  }
+  return ctx;
+};
